feat(feed): add audio-only /audio/podcast.xml alongside the full feed
Spotify rejects any feed carrying a video enclosure. Rather than strip
video from the existing feed (Apple and generic readers take it fine),
extract the builder to src/lib/podcast-feed.ts and serve two feeds:

  /audio/feed.xml     113 items — unchanged, video-only entries included
  /audio/podcast.xml   93 items — audioUrl only, for Spotify

Also label .m4a enclosures audio/mp4 rather than audio/mpeg in both:
the NLM takes are AAC in an MP4 container, and strict directory
validators check this.

Verified on the build: feed.xml unchanged at 113 items with its 20
video/mp4 enclosures; podcast.xml 93 items, zero video, zero zero-byte
lengths; each atom:link self-references its own path.
